fix: a warning gate, three compiler-only bugs, and two claims that were wrong
Ten warnings had accumulated in the repo's own code, none of them visible
because nothing failed on them. Fixed at the source, then pinned with
-DHOTPATH_WERROR=ON on every CI configure. -Werror stays off by default
locally: a newer compiler inventing a diagnostic should not stop work in
progress, but it must not be able to pile up unnoticed either.

Clean on all three front-ends now, 66/66 tests on each -- AppleClang 21/libc++,
GCC 13.3/libstdc++, Clang 18.1/libstdc++. Adds linux-clang to CI, because three
of the ten were invisible to the dev machine's compiler:

- strategy_eval computed skew_cost[m] and never read it, a leftover from a
  summary line that no longer exists. Nothing published was wrong -- the value
  is already printed inline -- but it is the shape of a dropped analysis.
- fuzz_campaign printed `seed + i * 0x9E37...ull` through PRIu64, which is
  "lu" on LP64 Linux while the expression is unsigned long long. Same width,
  different type, UB by the letter, and impossible to see on macOS where
  PRIu64 is "llu".
- The slow-consumer spin incremented a volatile lvalue, deprecated in C++20.
  clang flags ++k, gcc also flags k = k + 1; replaced with the repo's own
  do_not_optimize barrier, which is what the loop meant anyway.

scripts/linux_build.sh filtered its build output on "error:" only, so the one
environment that could see the gcc diagnostics was discarding them. Now shows
warnings too.

Also two documented claims that were simply wrong:

- The intrusive book's memmove volume was 2x too high in hybrid_book.hpp and
  BUILDLOG. The shipped level index is vector<int32_t>, so 2.9 GB per replay,
  not 5.7 GB -- that is the figure for the 8-byte entries of the *rejected*
  optimisation. intrusive_book.hpp:182 had it right and tape_stat measures
  2866 MB, so the repo contradicted itself in writing. Phase 2 records this
  exact bug as found and fixed; two instances survived. The error flattered
  the design being argued for.
- README "Known gaps" contradicted three of its own headline results: it
  claimed the fill model has "no latency on re-quotes" when an entire result
  sweeps exactly that from 0 to 10 ms, and said "four symbols" where the rest
  of the README says 25. Rewritten against what the code does, and extended
  with gaps that were true but unstated (no tail latencies; the RTL is
  simulated, not synthesised). Layout section brought back in sync too.

Bumps actions/checkout to v5, clearing the Node 20 deprecation notice.

+++ src/book_crossval.cpp
@@ -108,7 +108,8 @@ int main(int argc, char** argv) {
 
   std::printf("\n-- results --\n");
   std::printf("digest          : %016llx\n", (unsigned long long)digest.h);
-  std::printf("comparisons     : %zu\n", (tape.size() + check_every - 1) / check_every);
+  std::printf("comparisons     : %" PRIu64 "\n",
+              (static_cast<std::uint64_t>(tape.size()) + check_every - 1) / check_every);
   std::printf("divergences     : %" PRIu64 "\n", diverged);
   std::printf("final orders    : map=%zu intrusive=%zu flat=%zu hybrid=%zu\n",
               a.order_count(), b.order_count(), c.order_count(), d.order_count());
